[API] Simplify command submission
Command handling was in a weird state of old/prototype method of
handling commands inside traits and handling commands exclusively
in entity system actions.

I am settling to let traits validate whether commands are valid.
(We provide a way to inject a ICommandHandler in a default
trait implementation to avoid all the subclassing that kind of
extension normally brings.)

Traits are also implicitly responsible for cleaning up command
components on the actor (in case the system does not do it already).
The BasicCommandSystem is responsible for invoking that clean up
when necessary.